// src/app/models/enums.ts
export enum Status {
  active = "Active",
  inactive = "Inactive",
  pending = "Pending",
  blacklisted = "Blacklisted",
}

export enum Relationship {
  sister = "Sister",
  brother = "Brother",
  mother = "Mother",
  father = "Father",
}

export enum Gender {
  male = "Male",
  female = "Female",
}

export enum ResidenceType {
  apartment = "Apartment",
  duplex = "Duplex",
  villa = "Villa",
  townhouse = "TownHouse",
}

export enum MaritalStatus {
  single = "Single",
  relationship = "Relationship",
  married = "Married",
}